requirements.txt
django==1.10.3
Pillow
django-ckeditor
django-crispy-forms
